Test.Consensus.Cardano.Examples

Setup

codecConfig ∷ CardanoCodecConfig Crypto

Examples

exampleApplyTxErrWrongEraByron ∷ ApplyTxErr (CardanoBlock Crypto)

exampleApplyTxErrWrongEraShelley ∷ ApplyTxErr (CardanoBlock Crypto)

exampleEraMismatchByron ∷ MismatchEraInfo (CardanoEras Crypto)

Applying a Shelley thing to a Byron ledger

exampleEraMismatchShelley ∷ MismatchEraInfo (CardanoEras Crypto)

Applying a Byron thing to a Shelley ledger

exampleQueryAnytimeShelley ∷ SomeSecond BlockQuery (CardanoBlock Crypto)

exampleQueryEraMismatchByron ∷ SomeSecond BlockQuery (CardanoBlock Crypto)

exampleQueryEraMismatchShelley ∷ SomeSecond BlockQuery (CardanoBlock Crypto)

exampleResultAnytimeShelley ∷ SomeResult (CardanoBlock Crypto)

exampleResultEraMismatchByron ∷ SomeResult (CardanoBlock Crypto)

exampleResultEraMismatchShelley ∷ SomeResult (CardanoBlock Crypto)

examples ∷ Examples (CardanoBlock Crypto)

The examples: the examples from each individual era lifted in to CardanoBlock and the multi-era examples.

Orphan instances

Inject Serialised
In reality, an era tag would be prepended, but we're testing that the encoder doesn't care what the bytes are.

Instance details

Methods

inject ∷ ∀ x (xs ∷ [Type]). CanHardFork xs ⇒ Exactly xs Bound → Index xs x → Serialised x → Serialised (HardForkBlock xs)

Inject SomeResult

Instance details

Methods

inject ∷ ∀ x (xs ∷ [Type]). CanHardFork xs ⇒ Exactly xs Bound → Index xs x → SomeResult x → SomeResult (HardForkBlock xs)

Inject Examples

Instance details

Methods

inject ∷ ∀ x (xs ∷ [Type]). CanHardFork xs ⇒ Exactly xs Bound → Index xs x → Examples x → Examples (HardForkBlock xs)