Test.Consensus.Cardano.Generators

Description

Arbitrary instances intended for serialisation roundtrip tests for CardanoBlock and its related types.

Because the generated values are only used in serialisation roundtrip tests, they don't need to be valid blocks, transactions, etc.

We combine the Byron and Shelley-based instances defined elsewhere into Cardano instances by picking randomly from one of the eras.

Documentation

module Test.Consensus.Byron.Generators

Orphan instances

Arbitrary EraSummary

Instance details

Methods

arbitrary ∷ Gen EraSummary

shrink ∷ EraSummary → [EraSummary]

Arbitrary EraEnd

Instance details

Methods

arbitrary ∷ Gen EraEnd

shrink ∷ EraEnd → [EraEnd]

Arbitrary EraParams

Instance details

Methods

arbitrary ∷ Gen EraParams

shrink ∷ EraParams → [EraParams]

Arbitrary SafeZone

Instance details

Methods

arbitrary ∷ Gen SafeZone

shrink ∷ SafeZone → [SafeZone]

CardanoHardForkConstraints c ⇒ Arbitrary (HardForkNodeToNodeVersion (CardanoEras c))

Instance details

Methods

arbitrary ∷ Gen (HardForkNodeToNodeVersion (CardanoEras c))

shrink ∷ HardForkNodeToNodeVersion (CardanoEras c) → [HardForkNodeToNodeVersion (CardanoEras c)]

CardanoHardForkConstraints c ⇒ Arbitrary (HardForkNodeToClientVersion (CardanoEras c))

Instance details

Methods

arbitrary ∷ Gen (HardForkNodeToClientVersion (CardanoEras c))

shrink ∷ HardForkNodeToClientVersion (CardanoEras c) → [HardForkNodeToClientVersion (CardanoEras c)]

Arbitrary (BlockNodeToNodeVersion blk) ⇒ Arbitrary (EraNodeToNodeVersion blk)

Instance details

Methods

arbitrary ∷ Gen (EraNodeToNodeVersion blk)

shrink ∷ EraNodeToNodeVersion blk → [EraNodeToNodeVersion blk]

Arbitrary (BlockNodeToClientVersion blk) ⇒ Arbitrary (EraNodeToClientVersion blk)

Instance details

Methods

arbitrary ∷ Gen (EraNodeToClientVersion blk)

shrink ∷ EraNodeToClientVersion blk → [EraNodeToClientVersion blk]

Arbitrary (EraIndex (CardanoEras c))

Instance details

Methods

arbitrary ∷ Gen (EraIndex (CardanoEras c))

shrink ∷ EraIndex (CardanoEras c) → [EraIndex (CardanoEras c)]

Arbitrary (Interpreter (CardanoEras c))

Instance details

Methods

arbitrary ∷ Gen (Interpreter (CardanoEras c))

shrink ∷ Interpreter (CardanoEras c) → [Interpreter (CardanoEras c)]

(c ~ MockCryptoCompatByron, ShelleyBasedEra (ShelleyEra c)) ⇒ Arbitrary (AnnTip (CardanoBlock c))

Instance details

Methods

arbitrary ∷ Gen (AnnTip (CardanoBlock c))

shrink ∷ AnnTip (CardanoBlock c) → [AnnTip (CardanoBlock c)]

Arbitrary (CardanoBlock MockCryptoCompatByron)

Instance details

Methods

arbitrary ∷ Gen (CardanoBlock MockCryptoCompatByron)

shrink ∷ CardanoBlock MockCryptoCompatByron → [CardanoBlock MockCryptoCompatByron]

Arbitrary (CardanoHeader MockCryptoCompatByron)

Instance details

Methods

arbitrary ∷ Gen (CardanoHeader MockCryptoCompatByron)

shrink ∷ CardanoHeader MockCryptoCompatByron → [CardanoHeader MockCryptoCompatByron]

Arbitrary (Coherent (CardanoBlock MockCryptoCompatByron))

Instance details

Methods

arbitrary ∷ Gen (Coherent (CardanoBlock MockCryptoCompatByron))

shrink ∷ Coherent (CardanoBlock MockCryptoCompatByron) → [Coherent (CardanoBlock MockCryptoCompatByron)]

(CanMock (ShelleyEra c), CardanoHardForkConstraints c) ⇒ Arbitrary (OneEraHash (CardanoEras c))

Instance details

Methods

arbitrary ∷ Gen (OneEraHash (CardanoEras c))

shrink ∷ OneEraHash (CardanoEras c) → [OneEraHash (CardanoEras c)]

(Arbitrary a, SListI xs) ⇒ Arbitrary (NonEmpty xs a)

Instance details

Methods

arbitrary ∷ Gen (NonEmpty xs a)

shrink ∷ NonEmpty xs a → [NonEmpty xs a]

Arbitrary (Some QueryAnytime)

Instance details

Methods

arbitrary ∷ Gen (Some QueryAnytime)

shrink ∷ Some QueryAnytime → [Some QueryAnytime]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTxId c))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTxId c))

shrink ∷ WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTxId c) → [WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTxId c)]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTx c))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTx c))

shrink ∷ WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTx c) → [WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoGenTx c)]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoHeader c))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoHeader c))

shrink ∷ WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoHeader c) → [WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoHeader c)]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoBlock c))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoBlock c))

shrink ∷ WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoBlock c) → [WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (CardanoBlock c)]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (SomeSecond (NestedCtxt Header) (CardanoBlock c)))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (SomeSecond (NestedCtxt Header) (CardanoBlock c)))

shrink ∷ WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (SomeSecond (NestedCtxt Header) (CardanoBlock c)) → [WithVersion (HardForkNodeToNodeVersion (CardanoEras c)) (SomeSecond (NestedCtxt Header) (CardanoBlock c))]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeResult (CardanoBlock c)))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeResult (CardanoBlock c)))

shrink ∷ WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeResult (CardanoBlock c)) → [WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeResult (CardanoBlock c))]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeSecond BlockQuery (CardanoBlock c)))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeSecond BlockQuery (CardanoBlock c)))

shrink ∷ WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeSecond BlockQuery (CardanoBlock c)) → [WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (SomeSecond BlockQuery (CardanoBlock c))]

CardanoHardForkConstraints c ⇒ Arbitrary (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (Some (QueryHardFork (CardanoEras c))))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (Some (QueryHardFork (CardanoEras c))))

shrink ∷ WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (Some (QueryHardFork (CardanoEras c))) → [WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (Some (QueryHardFork (CardanoEras c)))]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoApplyTxErr c))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoApplyTxErr c))

shrink ∷ WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoApplyTxErr c) → [WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoApplyTxErr c)]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoGenTx c))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoGenTx c))

shrink ∷ WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoGenTx c) → [WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoGenTx c)]

c ~ MockCryptoCompatByron ⇒ Arbitrary (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoBlock c))

Instance details

Methods

arbitrary ∷ Gen (WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoBlock c))

shrink ∷ WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoBlock c) → [WithVersion (HardForkNodeToClientVersion (CardanoEras c)) (CardanoBlock c)]