Test.Consensus.Cardano.MockCrypto

Documentation

type BlockCompatByron h = ShelleyBlock (TPraosMockCryptoCompatByron h)

data TPraosMockCryptoCompatByron h

A mock replacement for TPraosStandardCrypto that is compatible with bootstrapping from Byron

This mocks more components than does TPraosMockCrypto. This prevents the cardano-ledger-specs generators from being re-used. Currently, this is not an obstacle for example in the Cardano ThreadNet tests.

NOTE: The Ouroboros.Consensus.Cardano.CanHardFork translation currently assumes that ADDRHASH has the same bit size as Byron address hashes (ie 224); that's why we use Blake2b_224 here.

NOTE: The shelley-spec-ledger package currently requires that @DSIGN ~ Ed25519DSIGN in order to use Byron bootstrap witnesses.