Test.ThreadNet.Infra.TwoEras

Description

Definitions used in ThreadNet tests that involve two eras.

Generators

data Partition

When and for how long the nodes are partitioned

The nodes are divided via message delays into two sub-networks by the parity of their CoreNodeId.

Constructors

Partition SlotNo NumSlots
the scheduled start slot and duration (which includes the start slot)

genNonce ∷ Gen Nonce

genPartition ∷ NumCoreNodes → NumSlots → SecurityParam → Gen Partition

Generate setupPartition

genTestConfig ∷ SecurityParam → (EpochSize, EpochSize) → Gen TestConfig

Generate a setupTestConfig relevant to the case where the first era (eg Byron) lasts for one epoch and the second era (eg Shelley) lasts for an interesting number of slots.

Era inspection

data ReachesEra2

Whether the test included second era blocks and (pre)reqs relevant to that

Note these fields are ordered alphabetically not semantically; see label_ReachesEra2.

Constructors

ReachesEra2

Fields

rsEra1Slots ∷ Prereq

enough slots in the first era to enable a block in the second era

rsPV ∷ Prereq

sufficient protocol version to enable a block in the second era

rsEra2Blocks ∷ Bool

blocks from the second era included in final chains

rsEra2Slots ∷ Requirement

enough slots in the second era to necessitate a block in the second era

activeSlotCoeff ∷ Rational

The active slot coefficient, f.

Some of these tests include epochs in the second era in which stakepools are actually leading. In that case, the k, d, and f parameters and the length of any scheduled network partitions need to be balanced so that Common Prefix violations (in particular, wedges) are extremely unlikely.

isFirstEraBlock ∷ HardForkBlock (era ': eras) → Bool

ledgerReachesEra2 ∷ ReachesEra2 → Bool

Is the update proposal adopted?

mkMessageDelay ∷ Partition → CalcMessageDelay blk

The temporary partition as a CalcMessageDelay

Calculates the delays that implement setupPartition.

numFirstEraEpochs ∷ Num a ⇒ a

The number of epochs in the first era in this test

All nodes join in slot 0, we generate the proposal in slot 0, we also generate the votes in slot 0, and the nodes are endorsing the proposal as of slot 0. Thus we expect that the first era will end after one epoch. Otherwise it would indicate some sort of protocol failure.

partitionExclusiveUpperBound ∷ Partition → SlotNo

secondEraOverlaySlots ∷ NumSlots → NumSlots → UnitInterval → EpochSize → Set SlotNo

All OBFT overlay slots in the second era.

shelleyEpochSize ∷ SecurityParam → Word64

Properties

label_ReachesEra2 ∷ ReachesEra2 → String

List the (pre)reqs in semantic order, followed by the observation

label_hadActiveNonOverlaySlots ∷ TestOutput (HardForkBlock (era ': eras)) → Set SlotNo → String

Whether there was a block forged in a non-overlay slot in the second era.

This event evidences that the stake pools were correctly created and delegated to.

prop_ReachesEra2 ∷ ReachesEra2 → Property

Checks if the observation satisfies the (pre)reqs

tabulateFinalIntersectionDepth ∷ SecurityParam → NumBlocks → String → Property → Property

tabulatePartitionDuration ∷ SecurityParam → Partition → Property → Property

tabulatePartitionPosition ∷ NumSlots → Partition → Bool → Property → Property