Test.Consensus.Cardano.MockCrypto

Documentation

data MockCryptoCompatByron

A replacement for MockCrypto that is compatible with bootstrapping from Byron.

The Ouroboros.Consensus.Cardano.CanHardFork translation requires that ADDRHASH has the same bit size as Byron address hashes (ie 224); that's why we use Blake2b_224 here.

Similarly, HASH has to have the same bit size as Byron header hashes (ie 256), that's why we use Blake2b_256 here.

The shelley-spec-ledger package currently requires that @DSIGN ~ Ed25519DSIGN in order to use Byron bootstrap witnesses.

We can still use mock KES and mock VRF.

Note that many Shelley generators are not instantiated to MockShelley but are constrained by CanMock era. ShelleyEra MockCryptoCompatByron satisfies this constraint, allowing us to reuse these generators for Cardano.